[client/rest]: promote shared NEM and Symbol rosetta utils
 problem: all rosetta utils are under /symbol but some can be used by nem
solution: move shared utils up a level
