+++ client/rest/src/plugins/rosetta/symbol/mempoolRoutes.js
@@ -20,14 +20,13 @@
  */
 
 import { OperationParser } from './OperationParser.js';
-import {
-	createLookupCurrencyFunction, rosettaPostRouteWithNetwork
-} from './rosettaUtils.js';
+import { createLookupCurrencyFunction, getBlockchainDescriptor } from './rosettaUtils.js';
 import MempoolResponse from '../openApi/model/MempoolResponse.js';
 import MempoolTransactionRequest from '../openApi/model/MempoolTransactionRequest.js';
 import MempoolTransactionResponse from '../openApi/model/MempoolTransactionResponse.js';
 import NetworkRequest from '../openApi/model/NetworkRequest.js';
 import TransactionIdentifier from '../openApi/model/TransactionIdentifier.js';
+import { rosettaPostRouteWithNetwork } from '../rosettaUtils.js';
 import { NetworkLocator } from 'symbol-sdk';
 import { Network } from 'symbol-sdk/symbol';
 
@@ -35,8 +34,8 @@ export default {
 	register: (server, db, services) => {
 		const PAGE_SIZE = 100;
 
-		const networkName = services.config.network.name;
-		const network = NetworkLocator.findByName(Network.NETWORKS, networkName);
+		const blockchainDescriptor = getBlockchainDescriptor(services.config);
+		const network = NetworkLocator.findByName(Network.NETWORKS, blockchainDescriptor.network);
 		const lookupCurrency = createLookupCurrencyFunction(services.proxy);
 		const parser = new OperationParser(network, {
 			includeFeeOperation: true,
@@ -45,7 +44,7 @@ export default {
 			resolveAddress: (address, transactionLocation) => services.proxy.resolveAddress(address, transactionLocation)
 		});
 
-		server.post('/mempool', rosettaPostRouteWithNetwork(networkName, NetworkRequest, async () => {
+		server.post('/mempool', rosettaPostRouteWithNetwork(blockchainDescriptor, NetworkRequest, async () => {
 			const transactions = await services.proxy.fetchAll('transactions/unconfirmed', PAGE_SIZE);
 
 			const response = new MempoolResponse();
@@ -53,12 +52,16 @@ export default {
 			return response;
 		}));
 
-		server.post('/mempool/transaction', rosettaPostRouteWithNetwork(networkName, MempoolTransactionRequest, async typedRequest => {
-			const transactionHash = typedRequest.transaction_identifier.hash;
-			const transaction = await services.proxy.fetch(`transactions/unconfirmed/${transactionHash}`);
-			const response = new MempoolTransactionResponse();
-			response.transaction = await parser.parseTransactionAsRosettaTransaction(transaction.transaction, transaction.meta);
-			return response;
-		}));
+		server.post('/mempool/transaction', rosettaPostRouteWithNetwork(
+			blockchainDescriptor,
+			MempoolTransactionRequest,
+			async typedRequest => {
+				const transactionHash = typedRequest.transaction_identifier.hash;
+				const transaction = await services.proxy.fetch(`transactions/unconfirmed/${transactionHash}`);
+				const response = new MempoolTransactionResponse();
+				response.transaction = await parser.parseTransactionAsRosettaTransaction(transaction.transaction, transaction.meta);
+				return response;
+			}
+		));
 	}
 };
