+++ client/rest/src/plugins/rosetta/symbol/rosettaUtils.js
@@ -19,89 +19,17 @@
  * along with Catapult.  If not, see <http://www.gnu.org/licenses/>.
  */
 
-import { sendJson } from '../../../routes/simpleSend.js';
 import Currency from '../openApi/model/Currency.js';
-import RosettaApiError from '../openApi/model/Error.js';
 
 /**
- * Error thrown when a rosetta endpoint encounters an error.
+ * Extracts blockchain descriptor from services configuration.
+ * @param {object} config Services configuration.
+ * @returns {object} Blockchain descriptor.
  */
-class RosettaError extends Error {
-	constructor(code, message, retriable) {
-		super(message);
-
-		this.apiError = new RosettaApiError(code, message, retriable);
-	}
-}
-
-/**
- * Factory for creating rosetta errors.
- */
-export class RosettaErrorFactory {
-	static get UNSUPPORTED_NETWORK() {
-		return new RosettaError(1, 'unsupported network', false);
-	}
-
-	static get UNSUPPORTED_CURVE() {
-		return new RosettaError(2, 'unsupported curve', false);
-	}
-
-	static get INVALID_PUBLIC_KEY() {
-		return new RosettaError(3, 'invalid public key', false);
-	}
-
-	static get INVALID_REQUEST_DATA() {
-		return new RosettaError(4, 'invalid request data', false);
-	}
-
-	static get CONNECTION_ERROR() {
-		return new RosettaError(5, 'unable to connect to network', true);
-	}
-
-	static get SYNC_DURING_OPERATION() {
-		return new RosettaError(6, 'sync was detected during operation', true);
-	}
-
-	static get NOT_SUPPORTED_ERROR() {
-		return new RosettaError(99, 'operation is not supported in rosetta', false);
-	}
-
-	static get INTERNAL_SERVER_ERROR() {
-		return new RosettaError(100, 'internal server error', false);
-	}
-}
-
-/**
- * Orchestrates a rosetta POST route by validating request data, including network, before calling user handler.
- * @param {string} networkName Network name.
- * @param {object} Request Type of request object.
- * @param {Function} handler User callback that is called with request data after validating request.
- * @returns {Function} Restify POST handler.
- */
-export const rosettaPostRouteWithNetwork = (networkName, Request, handler) => async (req, res, next) => {
-	const send = data => {
-		const sender = sendJson(res, next);
-		return data instanceof RosettaError ? sender(data.apiError, 500) : sender(data, undefined);
-	};
-
-	try {
-		Request.validateJSON(req.body);
-	} catch (err) {
-		return send(RosettaErrorFactory.INVALID_REQUEST_DATA);
-	}
-
-	const typedRequest = Request.constructFromObject(req.body);
-	const checkNetwork = networkIdentifier => ('Symbol' === networkIdentifier.blockchain && networkName === networkIdentifier.network);
-	if (!checkNetwork(typedRequest.network_identifier))
-		return send(RosettaErrorFactory.UNSUPPORTED_NETWORK);
-
-	try {
-		const response = await handler(typedRequest);
-		return send(response);
-	} catch (err) {
-		return send(err instanceof RosettaError ? err : RosettaErrorFactory.INTERNAL_SERVER_ERROR);
-	}
-};
+export const getBlockchainDescriptor = config => ({
+	blockchain: 'Symbol',
+	network: config.network.name
+});
 
 /**
  * Creates the lookup currency function used by the operation parser.
