+++ client/rest/test/plugins/rosetta/symbol/accountRoutes_spec.js
@@ -32,9 +32,9 @@ import Amount from '../../../../src/plugins/rosetta/openApi/model/Amount.js';
 import BlockIdentifier from '../../../../src/plugins/rosetta/openApi/model/BlockIdentifier.js';
 import Coin from '../../../../src/plugins/rosetta/openApi/model/Coin.js';
 import CoinIdentifier from '../../../../src/plugins/rosetta/openApi/model/CoinIdentifier.js';
+import { RosettaErrorFactory } from '../../../../src/plugins/rosetta/rosettaUtils.js';
 import { convertTransactionSdkJsonToRestJson } from '../../../../src/plugins/rosetta/symbol/OperationParser.js';
 import accountRoutes from '../../../../src/plugins/rosetta/symbol/accountRoutes.js';
-import { RosettaErrorFactory } from '../../../../src/plugins/rosetta/symbol/rosettaUtils.js';
 import { SymbolFacade, generateMosaicAliasId } from 'symbol-sdk/symbol';
 
 describe('account routes', () => {
